leaves/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import LeaveRequestRawView, LeaveRequestViewSet

router = DefaultRouter()
router.register(r'leave-requests', LeaveRequestViewSet, basename='leave-request')

urlpatterns = [
    # API Get data to show leave request form
    path(
        'leave-requests/<int:account_id>/<int:subject_id>/',
        LeaveRequestRawView.as_view(),
        name='get-data-to-leave-request'
    ),

    # API CRUD for LeaveRequest
    path('', include(router.urls)),
]