requirements.txt
appscript==1.3.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
cobble==0.1.4
constantly==23.10.4
cryptography==45.0.4
cssselect2==0.8.0
daphne==4.2.0
Django==5.2.3
django-cors-headers==4.7.0
django-crontab==0.7.1
django-crum==0.7.9
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.1
fonttools==4.59.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
lxml==6.0.1
mammoth==1.10.0
Markdown==3.9
MarkupSafe==3.0.2
msgpack==1.1.1
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyphen==0.17.2
python-docx==1.2.0
redis==6.2.0
reportlab==4.4.3
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1
gunicorn
psycopg2-binary